requirements.txt
ccxt==4.2.14
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
TA-Lib==0.4.28
python-dotenv==1.0.1
schedule==1.2.1
websocket-client==1.7.0
plotly==5.18.0
joblib==1.3.2
colorama==0.4.6
